+++ packages/design-system-web/src/components.d.ts
@@ -42,6 +42,8 @@ export namespace Components {
         "initValue"?: GeovClassSelectItem;
         "items"?: GeovClassSelectItem[];
         "loading"?: boolean;
+        "preferredItems"?: string[];
+        "uriPrefix": string;
     }
     interface GeovClassSelect {
         /**
@@ -282,6 +284,11 @@ export namespace Components {
          */
         "_ssrId"?: string;
         /**
+          * classUriPrefix Optional prefix for the class uris. Example: "http://geovistory.org/resource/" Default: "http://geovistory.org/resource/"
+          * @type {string}
+         */
+        "classUriPrefix"?: string;
+        /**
           * If true, the component will not render before the initial data is fetched
          */
         "fetchBeforeRender": boolean;
@@ -289,6 +296,10 @@ export namespace Components {
           * initialize the component with a given search string
          */
         "initSearchString"?: string;
+        /**
+          * preferredItems List of preferred items. If set, these items will be shown first in the list of items.  Example: ["c523", "c50"]
+         */
+        "preferredItems"?: string[];
         /**
           * sparqlEndpoint URL of the sparql endpoint
          */
@@ -540,6 +551,8 @@ declare namespace LocalJSX {
         "items"?: GeovClassSelectItem[];
         "loading"?: boolean;
         "onSelectionChanged"?: (event: GeovClassRadioGroupCustomEvent<GeovClassRadioGroupEvent>) => void;
+        "preferredItems"?: string[];
+        "uriPrefix"?: string;
     }
     interface GeovClassSelect {
         /**
@@ -784,6 +797,11 @@ declare namespace LocalJSX {
          */
         "_ssrId"?: string;
         /**
+          * classUriPrefix Optional prefix for the class uris. Example: "http://geovistory.org/resource/" Default: "http://geovistory.org/resource/"
+          * @type {string}
+         */
+        "classUriPrefix"?: string;
+        /**
           * If true, the component will not render before the initial data is fetched
          */
         "fetchBeforeRender"?: boolean;
@@ -791,6 +809,10 @@ declare namespace LocalJSX {
           * initialize the component with a given search string
          */
         "initSearchString"?: string;
+        /**
+          * preferredItems List of preferred items. If set, these items will be shown first in the list of items.  Example: ["c523", "c50"]
+         */
+        "preferredItems"?: string[];
         /**
           * sparqlEndpoint URL of the sparql endpoint
          */