+++ packages/design-system-react/CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 ## 1.10.1-pr-51.0 (2023-04-25)
+## 1.10.2-pr-56.0 (2023-05-03)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.10.1 (2023-05-02)
+
+**Note:** Version bump only for package @geovistory/design-system-react
+
+
+
+
+
+## 1.10.1-pr-48.0 (2023-04-19)
 
 **Note:** Version bump only for package @geovistory/design-system-react